+++ QuickGourmet/Repository/Search/ShopSearchRepository.swift
@@ -6,3 +6,34 @@
 //
 
 import Foundation
+
+class ShopSearchRepository: ShopSearchRepositoryInterface {
+    var requestString = ""
+
+    func fetchShopData(completion: @escaping ([Shop]) -> Void) {
+        guard let requestUrl = URL(string: requestString) else {
+            return
+        }
+        print("requestURL: \(requestUrl)")
+
+        URLSession.shared.dataTask(with: URLRequest(url: requestUrl)) { data, _, error in
+            guard let data = data else {
+                return
+            }
+
+            let decoder = JSONDecoder()
+            decoder.keyDecodingStrategy = .convertFromSnakeCase
+            do {
+                let searchResponseData = try decoder.decode(HotPepperResponse.self, from: data)
+                print("JSONDecode succeeded.")
+
+                DispatchQueue.main.async {
+                    completion(searchResponseData.results.shop)
+                }
+            } catch {
+                print("JSONDecode Failure Overview.: \(error.localizedDescription)")
+                print("Details of JSONDecode failure.: \(error)")
+            }
+        }.resume()
+    }
+}
